requirements.txt
asgiref==3.5.2
autopep8==1.6.0
beautifulsoup4==4.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.13
django-import-export==2.8.0
et-xmlfile==1.1.0
idna==3.3
lxml==4.9.0
MarkupPy==1.14
numpy==1.21.6
odfpy==1.4.1
openpyxl==3.0.10
pandas==1.3.5
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
toml==0.10.2
typing_extensions==4.2.0
urllib3==1.26.9
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
